patch for upcoming changes to gnupg
A new beta for GnuPG 2.1 has been released[0] and it may be production
ready soon. As part of the upcoming changes, the GPG_AGENT_INFO
environment variable will not be set anymore. The current gpg4me code
checks for this variable and gives a warning if it doesn't find it.
Given that it's not needed currently for 2.0 support if you use a fixed
socket, and the newest version will never set it, that check should
probably be changed.

I have included a small patch that changes the environment variable
check to use gpgme_engine_check_version to see if there is a PGP or CMS
engine available for use instead.

CustaiCo

[0] http://lists.gnupg.org/pipermail/gnupg-announce/2014q4/000357.html